requirements.txt
PyMuPDF>=1.23.0
nltk>=3.8
scikit-learn>=1.3
numpy>=1.24
streamlit>=1.30
matplotlib>=3.7
google-genai>=1.0
